move OwnedLiveEvent, etc. to types.rs
